backup: Separate operation on files from generating the content
Next changes related to GPG encryption are going to introduce a mixin
which is going to be inherited together with FileBackend. That mixin is
going to encrypt and decrypt data which is read from or written to
files. To make that easier, this change moves all operations on files to
separate methods which do not contain any logic about generating their
content.

Methods which generate content always return `bytes`. That is going to
make it easier for the GPG mixin to override them and encrypt/decrypt
those bytes, without interfering with file operations.

Signed-off-by: Michal Rostecki <mrostecki@mailfence.com>